Made the handling of types by eigency more implicit through cython and reduced overhead in conversions.pyx and conversions.pxd

+++ tests/run_tests.py
@@ -189,13 +189,13 @@ class TestEigency(unittest.TestCase):
 
     def test_function_type_int32(self):
         # C++ long - Note that this is the standard Python integer
-        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.long)
+        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.int32)
         mat_out = eigency_tests.function_type_int32(mat_in)
         assert_array_equal(mat_in, mat_out)
 
     def test_function_type_uint32(self):
         # C++ ulong
-        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.ulong)
+        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.uint32)
         mat_out = eigency_tests.function_type_uint32(mat_in)
         assert_array_equal(mat_in, mat_out)
